feat(#128): limiting search to user clients
If a user has a limited list of client numbers, the user should not be able to search for clients that are outside of this list. For IDIR users, they can search unrestricted. Most of the validation is already done in the backend service, allowing the legacy to have minimal checks.