feat(#144): adding sorting capabilities to table component
Added sorting capability to table component, to allow users to enable and handle sort. This includes the handling of the sortable property in the frontend headers object.
Along with that, the onSortChange function is how the table component will communicate sorting changes to the consumer, by providing the headers that were sorted. This is because as the table don't handle the actual data loading, it needs to notify the caller to handle the sort changes.
It also includes some refactor in the frontend to allow for a better code organization.